requirements.txt
absl-py
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
aistudio-sdk
annotated-types
anyio
astunparse
attrs
bce-python-sdk
blis
cachetools
catalogue
certifi
cffi
chardet
charset-normalizer
click
cloudpathlib
colorama
coloredlogs
colorlog
confection
contourpy
cryptography
cycler
cymem
datasets
dill
dnspython
easyocr
email_validator
et_xmlfile
filelock
flatbuffers
fonttools
frozenlist
fsspec
future
fuzzywuzzy
gast
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
google-pasta
googleapis-common-protos
grpcio
grpcio-status
h11
h5py
httpcore
httplib2
httpx
huggingface-hub
humanfriendly
idna
imageio
imagesize
Jinja2
joblib
jsonlines
keras
kiwisolver
langcodes
langdetect
language_data
lazy_loader
Levenshtein
libclang
llama2
lxml
marker-pdf
marisa-trie
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
ml_dtypes
modelscope
mpmath
multidict
multiprocess
murmurhash
namex
networkx
ninja
nltk
numpy
nvidia-ml-py
ollama
opencv-contrib-python
opencv-python
opencv-python-headless
openpyxl
opt-einsum
optree
outcome
packaging
paddleocr
paddlepaddle
paddlex
pandas
pdf2image
pdfminer.six
pdfplumber
phonenumbers
pillow
preshed
prettytable
propcache
proto-plus
protobuf
psutil
py-cpuinfo
pyarrow
pyasn1
pyasn1_modules
pyclipper
pycparser
pycryptodome
pydantic
pydantic_core
pyfiglet
Pygments
PyMuPDF
pynvml
pyparsing
pypdf
PyPDF2
pypdfium2
pyreadline3
PySocks
pytesseract
python-bidi
python-dateutil
python-docx
python-dotenv
python-Levenshtein
pytz
pywin32
PyYAML
RapidFuzz
regex
requests
rich
rsa
ruamel.yaml
ruamel.yaml.clib
scikit-image
scikit-learn
scipy
seaborn
selenium
setuptools
shapely
shellingham
six
smart_open
sniffio
sortedcontainers
spacy
spacy-legacy
spacy-loggers
srsly
sympy
tensorboard
tensorboard-data-server
tensorflow
termcolor
tf_keras
thinc
threadpoolctl
tifffile
tokenizers
tqdm
trio
trio-websocket
typer
typing-inspection
typing_extensions
tzdata
ujson
uritemplate
urllib3
wasabi
wcwidth
weasel
webdriver-manager
websocket-client
Werkzeug
wheel
wrapt
wsproto
xxhash
yarl